Added default values for RunOnVaultStart Split TimeSpanEx to separate file.
